Initial Commit
This commit has an initial folder structure setup. Much of this setup
is based on the Genesis Framework. I've included a few files from
Genesis to get an idea of what they have done in terms of
functionality, widgets, tools, short codes etc. I have found their file
structure and code to be very easy to work with. If nothing else, we
may find some of it useful or get some ideas of what to include.