Removed some default widgets and admin menus. Changed "Posts" to "News"
These can be updated, but for now, I thought we could do without some
default widgets and the links menu.

Also, updated "Posts" labels to "News". I was thinking this could be
more relative to what it will be used for News, Blog, Etc. This can be
updated as well.