Generate a debug notice when a deprecated function is called.
Most of the deprecated functions in 1.9.0 are admin hooks that shouldn't be used
elsewhere, but the definitions have been kept around to prevent fatal errors in
case someone did call them.

Calls to i18n functions in deprecated functions were also removed to keep them
out of the POT file.