Reorganize the file structure.
This structure more closely mimics PSR-4, although some classes will need to be
updated if we ever move to namespaces and a PSR-4 autoloader.

WP core's class naming convention also caused class names to become long,
especially in subdirectories, so it was difficult to view file names in tabs and
editor sidebars.