Introduce an address formatting class.
This attempts to format addresses based on the country the address is located
in. We don't store country codes, which makes looking up countries difficult due
to localization and other variations in the name, so it won't be perfect to
start, although it does allow the default format to be easily filtered.

Future iterations may map the site locale to an address format or ping an API
to look up country codes for countries that have been saved.

This replaces the current vcard functionality and slightly changes markup due
to differences in address formats, so some theme styles may need to be tweaked.